refactor: lift CLI logic to its own package
Cleanup main to make it easier to add new commands.
